Move types to its own file
